Simplify to a single image
We don't really need `_cli` and `_wsgi` separate images here, as they would start from the same base image and add the same virtualenv. So let's only build one image for the service for local usage.

Later on a `profiles_ci` (not `_cli`) can be added to contain tests and other development tools and make the main image lean.